requirements.txt
requests
feedparser
matplotlib
python-dotenv
openai
